requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
black==20.8b1
boto3==1.17.97
botocore==1.20.97
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-absoluteuri==1.3.0
django-apscheduler==0.5.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-on-heroku==1.1.2
django-restframework==0.0.1
django-storages==1.11.1
django-unixtimestampfield==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.9.0
flake8-django==1.1.1
flake8-print==4.0.0
google-auth==1.28.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
proto-plus==1.18.0
protobuf==3.15.6
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rope==0.18.0
rsa==4.7.2
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1